reorgnize sync primitives
